+++ SWP391_CareSkin_BE/Services/Interfaces/IPromotionService.cs
@@ -11,7 +11,7 @@ namespace SWP391_CareSkin_BE.Services.Interfaces
         Task<PromotionDTO> CreatePromotionAsync(PromotionCreateRequestDTO request);
         Task<PromotionDTO> UpdatePromotionAsync(int promotionId, PromotionUpdateRequestDTO request);
         Task<bool> DeletePromotionAsync(int promotionId);
-        Task<double> CalculateOrderDiscountAsync(int? promotionId, int customerId, double orderTotal);
+        Task<decimal> CalculateOrderDiscountAsync(int? promotionId, int customerId, decimal orderTotal);
         Task<List<PromotionDTO>> GetPromotionsForCustomerAsync(int customerId);
         Task<List<PromotionDTO>> GetPromotionsForProductAsync(int productId);
     }
